+++ apps/main-app/ts/features/authentication/login/cie/components/CieIdLoginWebView.tsx
@@ -24,15 +24,15 @@ import { IdpSuccessfulAuthentication } from "../../../common/components/IdpSucce
 import { AUTHENTICATION_ROUTES } from "../../../common/navigation/routes";
 import { loginFailure, loginSuccess } from "../../../common/store/actions";
 import { loggedInAuthSelector } from "../../../common/store/selectors";
-import { onLoginUriChanged } from "../../../common/utils";
-import { IdpCIE_ID } from "../../hooks/useNavigateToLoginMethod";
-import { LoadingOverlay } from "../shared/LoadingSpinnerOverlay";
+import { AUTH_LEVELS, onLoginUriChanged } from "../../../common/utils";
 import {
   CieIdLoginProps,
   defaultUserAgent,
   originSchemasWhiteList,
   WHITELISTED_DOMAINS
-} from "../shared/utils";
+} from "../../../common/utils/cie";
+import { IdpCIE_ID } from "../../hooks/useNavigateToLoginMethod";
+import { LoadingOverlay } from "../shared/LoadingSpinnerOverlay";
 import {
   getCieIdEnvironment,
   getCieIDLoginUri,
@@ -126,7 +126,7 @@ const CieIdLoginWebView = ({ spidLevel, isUat }: CieIdLoginProps) => {
         params: {
           errorCodeOrMessage: code || message,
           authMethod: "CIE_ID",
-          authLevel: "L2",
+          authLevel: AUTH_LEVELS.L2,
           params: { spidLevel, isUat }
         }
       });