+++ apps/main-app/ts/features/itwallet/statusList/utils/consts.ts
@@ -3,3 +3,5 @@ export const STORAGE_PREFIX = "@io.itwallet.statusList";
 export const STORAGE_KEY_LAST_CHECK_TIME = `${STORAGE_PREFIX}:lastCheckTime`;
 
 export const STORAGE_KEY_ITW_SPECS_VERSION = `${STORAGE_PREFIX}:itwSpecsVersion`;
+
+export const STORAGE_KEY_ITW_ENV = `${STORAGE_PREFIX}:itwEnv`;